pyproject.toml
[tool.poetry]
name = "statis-log"
version = "0.1.0"
description = "一个工程化的日志收集、分析和通知工具"
authors = ["Jokerkern <852946650@qq.com>"]
readme = "README.md"
packages = [{include = "statis_log"}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pyyaml = ">=5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^4.1.0"
isort = "^5.12.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
statis-log = "statis_log.cli:main"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
